+++ .gitignore
@@ -1,2 +1,3 @@
+coverage
+dist
 node_modules
-dist/
