+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pixijs-actions",
-  "version": "1.1.2",
+  "version": "1.1.3",
   "author": "Reece Como <reece.como@gmail.com>",
   "authors": [
     "Reece Como <reece.como@gmail.com>",
