+++ src/actions/delay/DelayAction.ts
@@ -1,0 +1,11 @@
+
+import { Action } from '../../lib/Action';
+
+export class DelayAction extends Action {
+  protected onTick(): void {
+  }
+
+  public reversed(): Action {
+    return this;
+  }
+}
