+++ src/actions/display-object/index.ts
@@ -1,0 +1,3 @@
+export * from './RemoveFromParentAction';
+export * from './RunOnChildWithNameAction';
+export * from './SetVisibleAction';
