+++ src/actions/follow-path/index.ts
@@ -1,0 +1,1 @@
+export * from './FollowPathAction';
