+++ src/actions/index.ts
@@ -1,0 +1,10 @@
+export * from './chainable';
+export * from './custom';
+export * from './delay';
+export * from './display-object';
+export * from './follow-path';
+export * from './move';
+export * from './rotate';
+export * from './scale';
+export * from './speed';
+export * from './transparency';
