+++ src/actions/move/index.ts
@@ -1,0 +1,2 @@
+export * from './MoveByAction';
+export * from './MoveToAction';
