+++ src/actions/rotate/index.ts
@@ -1,0 +1,2 @@
+export * from './RotateByAction';
+export * from './RotateToAction';
