+++ src/actions/scale/index.ts
@@ -1,0 +1,3 @@
+export * from './ScaleByAction';
+export * from './ScaleToAction';
+export * from './ScaleToSizeAction';
