+++ src/actions/speed/index.ts
@@ -1,0 +1,2 @@
+export * from './SpeedByAction';
+export * from './SpeedToAction';
