+++ src/actions/transparency/index.ts
@@ -1,0 +1,4 @@
+export * from './FadeByAction';
+export * from './FadeInAction';
+export * from './FadeOutAction';
+export * from './FadeToAction';
