+++ tsconfig.json
@@ -5,7 +5,7 @@
     "noImplicitAny": true,
     "module": "es6",
     "target": "es6",
-    "allowJs": true,
+    "allowJs": false,
     "moduleResolution": "node",
     "downlevelIteration": true,
     "declaration": true,
@@ -14,6 +14,6 @@
     "baseUrl": "./",
     "skipLibCheck": true
   },
-  "include": ["src/**/*"],
-  "exclude": ["src/test/**/*"]
+  "include": ["src/*", "src/**/*", "src/**/**/*"],
+  "exclude": ["src/__tests__/**/*"]
 }
